+++ starsky/starsky.foundation.platform/Architecture/OperatingSystemHelper.cs
@@ -7,7 +7,7 @@ namespace starsky.foundation.platform.Architecture;
 
 public static class OperatingSystemHelper
 {
-	public delegate bool IsOsPlatformDelegate(OSPlatform osPlatform);
+	internal delegate bool IsOsPlatformDelegate(OSPlatform osPlatform);
 
 	public static OSPlatform GetPlatform()
 	{